- Add Support of the MongoDB Java Driver.   modified:   storage/connect/CMakeLists.txt   modified:   storage/connect/JavaWrappers.jar   modified:   storage/connect/colblk.h   modified:   storage/connect/filter.cpp   modified:   storage/connect/filter.h   modified:   storage/connect/ha_connect.cc   modified:   storage/connect/ha_connect.h   modified:   storage/connect/jdbccat.h   modified:   storage/connect/jdbconn.cpp   modified:   storage/connect/jdbconn.h   modified:   storage/connect/mongofam.cpp   modified:   storage/connect/mongofam.h   modified:   storage/connect/mycat.cc   modified:   storage/connect/mycat.h   modified:   storage/connect/tabext.h   modified:   storage/connect/tabjdbc.cpp   modified:   storage/connect/tabjdbc.h   modified:   storage/connect/tabjson.cpp   modified:   storage/connect/tabjson.h   modified:   storage/connect/tabmgo.cpp   modified:   storage/connect/tabmgo.h   created:    storage/connect/Mongo2Interface.java   created:    storage/connect/Mongo3Interface.java   created:    storage/connect/cmgoconn.cpp   created:    storage/connect/cmgoconn.h   created:    storage/connect/javaconn.cpp   created:    storage/connect/javaconn.h   created:    storage/connect/jmgfam.cpp   created:    storage/connect/jmgfam.h   created:    storage/connect/jmgoconn.cpp   created:    storage/connect/jmgoconn.h   created:    storage/connect/mongo.cpp   created:    storage/connect/mongo.h   created:    storage/connect/tabjmg.cpp   created:    storage/connect/tabjmg.h
- tdbp not initialized when catched exception
in CntGetTDB (connect.cc line 188)
  modified:   storage/connect/connect.h

- CheckCleanup should sometimes doing cleanup on pure info
Sometimes MariaDB loops on info to get the size of all tables in a database.
This can sometimes fail by exhausted memory.
CheckCleanup now have a force boolean parameter (defaulting to false)
  modified:   storage/connect/ha_connect.cc
  modified:   storage/connect/user_connect.cc
  modified:   storage/connect/user_connect.h

Change the copyright of some source files
  modified:   storage/connect/connect.cc
  modified:   storage/connect/connect.h
  modified:   storage/connect/engmsg.h
  modified:   storage/connect/global.h
  modified:   storage/connect/ha_connect.cc
  modified:   storage/connect/ha_connect.h
  modified:   storage/connect/msgid.h
  modified:   storage/connect/mycat.cc
  modified:   storage/connect/mycat.h
  modified:   storage/connect/os.h
  modified:   storage/connect/osutil.c
  modified:   storage/connect/osutil.h
  modified:   storage/connect/user_connect.cc
  modified:   storage/connect/user_connect.h

+++ storage/connect/user_connect.cc
@@ -1,4 +1,4 @@
-/* Copyright (C) Olivier Bertrand 2004 - 2015
+/* Copyright (C) MariaDB Corporation Ab
 
   This program is free software; you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
@@ -24,7 +24,7 @@
   that is a connection with its personnal memory allocation.
 
   @note
-
+	Author Olivier Bertrand
 */
 
 /****************************************************************************/
@@ -144,9 +144,9 @@ void user_connect::SetHandler(ha_connect *hc)
 /****************************************************************************/
 /*  Check whether we begin a new query and if so cleanup the previous one.  */
 /****************************************************************************/
-bool user_connect::CheckCleanup(void)
+bool user_connect::CheckCleanup(bool force)
 {
-  if (thdp->query_id > last_query_id) {
+  if (thdp->query_id > last_query_id || force) {
     uint worksize= GetWorkSize();
 
     PlugCleanup(g, true);
@@ -171,7 +171,7 @@ bool user_connect::CheckCleanup(void)
     g->Mrr = 0;
     last_query_id= thdp->query_id;
 
-    if (trace)
+    if (trace && !force)
       printf("=====> Begin new query %llu\n", last_query_id);
 
     return true;
